requirements.txt
flask
torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
clip @ git+https://github.com/openai/CLIP.git
Pillow
matplotlib
tqdm
